% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_starting_pos.R
\name{find_starting_pos}
\alias{find_starting_pos}
\title{test}
\usage{
find_starting_pos(nodes, lon_start, lat_start, n)
}
\arguments{
\item{nodes}{test}

\item{lon_start}{test}

\item{lat_start}{test}

\item{n}{test}
}
\value{
test
}
\description{
test
}
