% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{get_ages}
\alias{get_ages}
\title{get n ages according to age distribution at time}
\usage{
get_ages(t, n)
}
\arguments{
\item{t}{double time}

\item{n}{integer amount}
}
\value{
vector of ages
}
\description{
get n ages according to age distribution at time
}
