% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{get_attribute}
\alias{get_attribute}
\title{get n sexes according to sex distribution at time}
\usage{
get_attribute(t = NA, n = NA, distribution_function, range)
}
\arguments{
\item{t}{double time}

\item{n}{integer amount}
}
\value{
vector of sexes
}
\description{
get n sexes according to sex distribution at time
}
