% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{get_sexes}
\alias{get_sexes}
\title{get n sexes according to sex distribution at time}
\usage{
get_sexes(t, n)
}
\arguments{
\item{t}{double time}

\item{n}{integer amount}
}
\value{
vector of sexes
}
\description{
get n sexes according to sex distribution at time
}
