% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_attributes.R
\name{get_units}
\alias{get_units}
\title{get n units according to unit distribution at time}
\usage{
get_units(t, n, unit_vector)
}
\arguments{
\item{t}{double time}

\item{n}{integer amount}
}
\value{
vector of sexes
}
\description{
get n units according to unit distribution at time
}
