% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphml.R
\name{graphread}
\alias{graphread}
\title{string->igraph}
\usage{
graphread(inputstring, format = "graphml")
}
\arguments{
\item{inputstring}{string that describes a graph}

\item{format}{format of the string}
}
\value{
igraph object
}
\description{
create igraph object from string
}
