% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphml.R
\name{graphwrite}
\alias{graphwrite}
\title{igraph->string}
\usage{
graphwrite(igraphobj, format = "graphml")
}
\arguments{
\item{igraphobj}{igraph object}

\item{format}{format of the string}
}
\value{
string
}
\description{
create string from igraph object
}
