% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexify.R
\name{hexify}
\alias{hexify}
\title{test}
\usage{
hexify(areapath, clippath, hexproj, bufferwidth, hexcellsize)
}
\arguments{
\item{areapath}{test}

\item{clippath}{test}

\item{hexproj}{test}

\item{bufferwidth}{test}

\item{hexcellsize}{test}
}
\value{
test
}
\description{
test
}
