% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_ideas_world.R
\name{link_ideas_world}
\alias{link_ideas_world}
\title{add input graph vertex coordinates to model results}
\usage{
link_ideas_world(idea_exp, world_graph)
}
\arguments{
\item{idea_exp}{igraph graph object}

\item{world_graph}{igraph graph object}
}
\value{
data.frame
}
\description{
add input graph vertex coordinates to model results
}
