% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\docType{class}
\name{model_builder-class}
\alias{model_builder-class}
\title{model_builder class}
\description{
class to store values that should be passed to the model
}
\section{Slots}{

\describe{
\item{\code{networkland_env}}{string: Description of input graph in graphml format}

\item{\code{number_iterations}}{integer: Number of model iterations}

\item{\code{initial_idea_starting_positions}}{numeric vector: ids of nodes where the initial idea starts}
}}

