% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_devel.R
\name{plot_devel}
\alias{plot_devel}
\title{plot function - network development}
\usage{
plot_devel(graph, runres, store = FALSE, path = "../modevel/", world, hex)
}
\arguments{
\item{graph}{igraph graph object}

\item{runres}{runres dataframe with columns timestep, ideas, vertices}

\item{store}{switch to decide, if the plots should be stored in the
file system. Default: FALSE}

\item{path}{directory in file system, where the plots are stored.
\bold{Other files in this directory are deleted.}
Default: "../modevel/"}

\item{world}{test}

\item{hex}{test}
}
\value{
NULL - just called for side effects
}
\description{
Creates graph plots for a model run.
}
