% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_state}
\alias{plot_state}
\title{plot function - add ideas to world plot}
\usage{
plot_state(worldplot, states, state_id)
}
\arguments{
\item{worldplot}{ggplot world plot}

\item{states}{result df of link_ideas_world}

\item{state_id}{integer. id of state to plot}
}
\value{
resplot
}
\description{
plot function - add ideas to world plot
}
