% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_world.R
\name{plot_world}
\alias{plot_world}
\title{plot function - world}
\usage{
plot_world(graph, world_polygon, hex = NULL, plotedges = FALSE)
}
\arguments{
\item{graph}{igraph graph object}

\item{world_polygon}{polygon data.frame describing the land outline}

\item{hex}{polygon data.frame describing the hex raster}

\item{plotedges}{should the graph edges be plotted?}
}
\value{
resplot
}
\description{
plot function - world
}
