% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_growth_simulator.R
\name{simulate_growth}
\alias{simulate_growth}
\title{simulate growth of a population}
\usage{
simulate_growth(humans, time = 1:2000)
}
\arguments{
\item{humans}{tibble input population}

\item{time}{vector time steps}
}
\value{
tibble output population
}
\description{
simulate growth of a population
}
