% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_world.R
\name{plot_world}
\alias{plot_world}
\title{plot function - world}
\usage{
plot_world(graph, world, hex = NULL, plotedges = FALSE)
}
\arguments{
\item{graph}{igraph graph object}

\item{world}{test}

\item{hex}{test}

\item{plotedges}{test}
}
\value{
resplot
}
\description{
Creates graph plots for a model run.
}
