% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_world.R
\name{plot_world}
\alias{plot_world}
\title{plot function - world}
\usage{
plot_world(graph, world_polygon, regions = NULL, nodes = NULL,
  plotedges = TRUE)
}
\arguments{
\item{graph}{igraph graph object}

\item{world_polygon}{polygon data.frame describing the land outline}

\item{regions}{polygon data.frame describing the regions}

\item{nodes}{data.frame with info about the nodes}

\item{plotedges}{should the graph edges be plotted?}
}
\value{
resplot
}
\description{
plot function - world
}
