% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{run model}
\usage{
run(popdevel, steps = 100L)
}
\arguments{
\item{popdevel}{test}

\item{steps}{test}
}
\value{
test
}
\description{
test
}
\details{
test
}
\examples{
f <- function(size,birthrate,deathrate) {
  res <- size + size*birthrate*2 - size*deathrate*2
  return(res)
}

run(f)

}

