% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{run model}
\usage{
run(model_builder)
}
\arguments{
\item{model_builder}{model_builder-object with parameters}
}
\value{
List with different diagnostic values
}
\description{
General run method. Something like the main function of a usual
c++-program. Runs model with input parameters and gives results
directly back to R.
}
